chore(config): move storybook config inside src dir
