+++ src/storybook/preview-head.html
@@ -1,0 +1,8 @@
+<link rel="preconnect" href="https://fonts.googleapis.com" />
+<link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
+<link rel="preload" href="https://asset.fithub.id/fonts/fithub-icon/FitSystemIcons.woff2" as="font" crossorigin />
+<link rel="preload" href="https://asset.fithub.id/fonts/fithub-icon/FitSystemIcons.woff" as="font" crossorigin />
+<link rel="preload" href="https://asset.fithub.id/fonts/fithub-icon/style.min.css" as="style" />
+<link rel="stylesheet" href="https://asset.fithub.id/fonts/fithub-icon/style.min.css" />
+<link href="https://fonts.googleapis.com/css2?family=Inter:wght@400;500;600;700&display=swap" as="style" rel="preload" />
+<link href="https://fonts.googleapis.com/css2?family=Inter:wght@400;500;600;700&display=swap" rel="stylesheet" />
